Creation of the recipe app
Main menu- provides functions that the user can interact with
The user can create a recipe.
The User can search a recipe from its ingredients
The User can update a created function.
The User can delete a recipe.
The User can view all the recipes.
And the user quits the app when typing quit.